fpart: exclude files earlier (fixes GH issue #14)
With that change, and when using options -z and -x together, a directory will
be considered as empty (and thus packed) if it has been drained because of
option -x excluding all its contents.

Also, excluded files will no more account for a directory size when options -D
or -E are used.

Reported by: Alex Hunsley through Github issue #14 (thanks!)