Debugging (POLA): display '<empty>' when p->fts_name is a 0-length string
This is observed when a path supplied to fts_open(3) is a directory explicitly
ending with a '/' character. In that case, fts(3) returns an empty fts_name for
the root dir itself.

To reproduce, build fpart with debugging enabled, then:
 $ mkdir /tmp/root
 $ fpart -n 1 -v /tmp/root  2>&1 | grep '<empty>' => no match
 $ fpart -n 1 -v /tmp/root/ 2>&1 | grep '<empty>' => match

Filters are affected by that behaviour because they use fts_name. This is
expected as /tmp/root/ means the *contents* of /tmp/root/, not the directory
itself.