Fpsync: force non-recursive rsync(1) in file-based mode
Fix long-standing error where rsync(1) was called with a useless (and even
*wrong*) recursive (-r) option for file-based syncs.

In reality, that bug should not be triggered because in file-based mode, only
empty/unreachable directories appear in file lists, so there is usually nothing
to be copied recursively. But with the appearance of replay options, replaying
a run containing such directories could lead to copying them recursively if
they have been populated (or their access has been fixed) in the meantime.

To avoid such situations, let's force non-recursive rsync calls as we are
working with leaf elements *only*.