fpsync: Run resuming - Rework commit 686edde
To improve efficiency, do not actively track jobs' activity anymore but
just mark jobs as started and exited. To mark a job as started, its job
file is touched, to mark it as exited, its timestamps are reset to
partition 1's [1]. That permits rescheduling only necessary jobs and
not last OPT_RJOBS jobs (worst case scenario).

The drawback is that jobs killed by an external program will currently
be considered as "normally" finished and not rescheduled when resuming
a run. This is probably acceptable as the normal workflow is to kill
jobs from fpsync itself. When manually (externally) killing jobs, we
may (?) not want them to remain stuck to the resumable state. Anyway,
the killed state will be reflected in the return code so one can always
replay the entire run if necessary.

Also, fix counters erroneously including killed jobs.

[1] Another fix would have been to reset timestamps based on
corresponding job's partition, but that can lead to missing
long-running jobs when resuming a run: when fpart is still crawling and
generates a short-running-already-terminated job just after a
long-running-not-terminated one.