fpsync: Work-around hardcoded /bin/sh not providing 'pipefail' option on Debian
On Debian, Dash does not support the 'pipefail' option, necessary to detect job
errors when using piped tools (cpio, pax, tar). Fpsync will now display a
warning and disable the option when the shell does not support it.

Introduce an OPT_JOBSSHELL (internal-only) option to set the shell to use within
generated jobs, allowing to select a shell supporting the 'pipefail' option,
instead of the previously-hardcoded /bin/sh one.

That shell must exist locally and remotely (on workers) as well as support -c
and -s options.

This is a followup to commit 3eff9807.